Mostrando as chaves dos documentos somente quando emitido em contingencia
