+++ tests/bootstrap.php
@@ -1,0 +1,4 @@
+<?php
+require 'vendor/autoload.php';
+
+define('TEST_FIXTURES', __DIR__ . '/fixtures/');
